requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-widget-tweaks==1.5.0
dnspython==2.7.0
gunicorn==23.0.0
idna==3.10
mongoengine==0.29.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pymongo==3.11.4
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
